Skip Docker image builds for Greenkeeper branches
We have found that builds for Greenkeeper branches fail because the
branch names have characters in them that are not valid in a Docker
image tag. Fortunately, we also don't really want these branches to
result in CI built Docker images anyway, so an easy workaround is to
skip building Docker images on those branches.

Connects https://github.com/pelias/dockerfiles/issues/21