Put verbose point logging behind a config flag
This importer currently logs _every_ record imported as well as its name
and centroid. While useful for debugging, it's not needed all the time,
even at the `verbose` logger level.

By putting this option behind a `pelias.json` config flag (defaulting to
off) we can have it when we need it, and save over 10GB in each of our
build logs when we don't.