feat(docs): Rewrite configuration in a more readable format
I've found that GitHub does not do a good job of rendering our standard
table layouts for configuration values. The descriptions often end up
broken across many short lines that are hard to read.

In the case of this importer's docs, the table was broken, and there was
a section of config documentation below it. The configuration docs have
been rewritten in a way that is hopefully more readable and consistent.