feat(schema): Use layer instead of Elasticsearch _type
After https://github.com/pelias/model/pull/95, and in anticipation of
Elasticsearch 6 support, all use of the `_type` property on
Elasticsearch documents has to be avoided.

This updates the end to end test expectations with the new single
document type, and updates some lingering references in the code to the
`_type` field.