feat(popularity): Add configuration option to control venue removal
As described in https://github.com/pelias/openstreetmap/issues/537, the
default set in https://github.com/pelias/openstreetmap/pull/493, where
all venues that have a calculated popularity below `0` are not imported,
is a bit strict.

This adds a config flag, `imports.openstreetmap.removeDisusedVenues`
that controls whether or not that behavior is activated.

In addition, when enabled, a `warning` is displayed for each removed
record.